requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
llama-cpp-python>=0.2.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
